`hw-rust` update to align with `xapp-frame-rust`
Updated `hw-rust` app to align with `0.3.0-dev` Rust XApp Framework
(`xapp-frame-rust`) -
1. Using the new `from_config` API
2. Removed the `Ready/Live` server (now supported in the framework)

patchset-2:

Fixed the `hw-rust` App to use the new API. Since the `Processor` thread
is no longer used, the corresponding changes are required to be done in
the `hw-rust`. This commit does that.

patchset-3:

Updated the Dockerfile to use `1-bullseye` image to build using latest
Rust release

Issue-Id: RICAPP-232
Change-Id: I742a4a517df557ea5daa462d24cdeb26bd2d9ca0
Signed-off-by: Abhijit Gadgil <gabhijit@iitbombay.org>